move ProfilerResultStorage to own module to resolve import
